Fix fixed point parsing
The fraction part was overflowing, use 32 bit variable to prevent that.
Can also use division and shifting in stages to avoid the issue without
the 32 bit variable but that would be an optimization for later.

+++ stm8/main.c
@@ -224,7 +224,7 @@ uint16_t parse_fixed_point(uint8_t *s)
 {
 	uint8_t *t = s;
 	uint16_t val = 0;
-	uint16_t fraction_digits = 0;
+	uint32_t fraction_digits = 0;
 	uint16_t whole_digits = 0;
 	uint16_t fraction_factor = 1;
